Restructures front-end routing for Reitit
This is a pretty major re-factor.  The goal was to make the page more
SPA-like while reflecting browser navigation in the URL.  It took a lot
of google searches to find a workable design pattern for reitit +
re-frame that was front-end specific.

This commit is a WIP - as the routing works now, but I need to do some
code cleanup.

TODO:
1. Move all routing logic to its own namespace
2. Move all views to their own namespaces
3. Add working hrefs for specific blog posts in Posts tab