Resolve HTTPS redirect in prod
I do not want https redirect to happen locally.  The middleware in
luminus is already set up in the correct order, but the defaults are all
set to false for https redirecting.

Here is a stackoverflow link with more details:
https://stackoverflow.com/questions/30689577/clojure-wrap-ssl-redirect-on-heroku

In env.clj, there are settings for production.

To read from the config file (like in core.clj):
`(update :port #(or (-> env :options :port) %))`

Let's hope this works?